# app/routers/produits.py
from fastapi import APIRouter, Depends, HTTPException
from fastapi import Query
from typing import Optional
from sqlalchemy.orm import Session
from app.db import get_session
from app.schemas import ProduitUpdate
from app.services.produit_service import rechercher_produits_avances
from app.services.produit_service import (
    afficher_tout_le_stock,
    rechercher_produit,
    ajouter_produit,
    modifier_produit,
)

router = APIRouter(prefix="/api/produits", tags=["Produits"])


@router.get(
    "/filtrer",
    summary="Filtrer, trier et paginer les produits",
    description="Permet de filtrer les produits par catégorie, trier et paginer les résultats.",
)
def filtrer_produits(
    page: int = Query(1, ge=1),
    size: int = Query(10, ge=1, le=100),
    sort: str = Query("nom,asc"),
    categorie: Optional[str] = None,
    session: Session = Depends(get_session),
):
    return rechercher_produits_avances(page, size, sort, categorie, session)


@router.get(
    "/",
    summary="Lister tous les produits",
    description="Retourne tous les produits enregistrés dans tous les magasins.",
)
def get_all(session: Session = Depends(get_session)):
    return afficher_tout_le_stock(session)


@router.get(
    "/recherche",
    summary="Rechercher un produit",
    description="Recherche un produit par nom, catégorie ou ID numérique.",
)
def rechercher(critere: str, session: Session = Depends(get_session)):
    return rechercher_produit(critere, session)


@router.post(
    "/",
    summary="Ajouter un nouveau produit",
    description="Ajoute un nouveau produit dans un magasin donné. Le magasin est identifié par `magasin_id`.",
)
def ajouter(
    nom: str,
    categorie: str,
    prix: float,
    quantite_stock: int,
    magasin_id: int,
    session: Session = Depends(get_session),
):
    return ajouter_produit(nom, categorie, prix, quantite_stock, magasin_id, session)


@router.put(
    "/{produit_id}",
    summary="Mettre à jour un produit",
    description="Modifie un ou plusieurs champs (nom, prix, quantité...) d’un produit existant.",
)
def update_produit(
    produit_id: int, update_data: ProduitUpdate, session: Session = Depends(get_session)
):
    updated = modifier_produit(
        produit_id, update_data.dict(exclude_unset=True), session
    )
    if not updated:
        raise HTTPException(status_code=404, detail="Produit non trouvé")
    return updated