# docs/LAB3/openapi.json
{"openapi":"3.1.0","info":{"title":"Système de caisse","description":"API LOG430 avec token statique de sécurité","version":"1.0.0"},"paths":{"/api/produits/":{"get":{"tags":["Produits"],"summary":"Lister tous les produits","description":"Retourne tous les produits enregistrés dans tous les magasins.","operationId":"get_all_api_produits__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"APIKeyHeader":[]}]},"post":{"tags":["Produits"],"summary":"Ajouter un nouveau produit","description":"Ajoute un nouveau produit dans un magasin donné. Le magasin est identifié par `magasin_id`.","operationId":"ajouter_api_produits__post","parameters":[{"name":"nom","in":"query","required":true,"schema":{"type":"string","title":"Nom"}},{"name":"categorie","in":"query","required":true,"schema":{"type":"string","title":"Categorie"}},{"name":"prix","in":"query","required":true,"schema":{"type":"number","title":"Prix"}},{"name":"quantite_stock","in":"query","required":true,"schema":{"type":"integer","title":"Quantite Stock"}},{"name":"magasin_id","in":"query","required":true,"schema":{"type":"integer","title":"Magasin Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyHeader":[]}]}},"/api/produits/recherche":{"get":{"tags":["Produits"],"summary":"Rechercher un produit","description":"Recherche un produit par nom, catégorie ou ID numérique.","operationId":"rechercher_api_produits_recherche_get","parameters":[{"name":"critere","in":"query","required":true,"schema":{"type":"string","title":"Critere"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyHeader":[]}]}},"/api/produits/{produit_id}":{"put":{"tags":["Produits"],"summary":"Mettre à jour un produit","description":"Modifie un ou plusieurs champs (nom, prix, quantité...) d’un produit existant.","operationId":"update_produit_api_produits__produit_id__put","parameters":[{"name":"produit_id","in":"path","required":true,"schema":{"type":"integer","title":"Produit Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/ProduitUpdate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyHeader":[]}]}},"/api/ventes/":{"get":{"tags":["Ventes"],"summary":"Lister toutes les ventes","description":"Retourne la liste complète des ventes enregistrées dans tous les magasins.","operationId":"get_ventes_api_ventes__get","parameters":[{"name":"x-token","in":"header","required":true,"schema":{"type":"string","title":"X-Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/VenteOut"},"title":"Response Get Ventes Api Ventes  Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyHeader":[]}]},"post":{"tags":["Ventes"],"summary":"Enregistrer une nouvelle vente","description":"Crée une vente dans un magasin donné.\n\nBody attendu :\n```json\n{\n  \"magasin_id\": 1,\n  \"date\": \"2025-07-02\",  // optionnel\n  \"panier\": [\n    { \"produit_id\": 1, \"quantite\": 2 },\n    { \"produit_id\": 3, \"quantite\": 1 }\n  ]\n}","operationId":"creer_vente_api_ventes__post","parameters":[{"name":"x-token","in":"header","required":true,"schema":{"type":"string","title":"X-Token"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"type":"object","additionalProperties":true,"title":"Data"}}}},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyHeader":[]}]}},"/api/ventes/rapport":{"get":{"tags":["Ventes"],"summary":"Générer un rapport global des ventes","description":"Retourne un rapport contenant le chiffre d’affaires global par produit et par magasin.","operationId":"rapport_global_api_ventes_rapport_get","parameters":[{"name":"x-token","in":"header","required":true,"schema":{"type":"string","title":"X-Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PerformanceGlobaleDTO"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyHeader":[]}]}},"/api/ventes/{vente_id}":{"delete":{"tags":["Ventes"],"summary":"Annuler une vente","description":"Annule une vente en la supprimant du système si elle n’a pas déjà été annulée.","operationId":"supprimer_vente_api_ventes__vente_id__delete","parameters":[{"name":"vente_id","in":"path","required":true,"schema":{"type":"integer","title":"Vente Id"}},{"name":"x-token","in":"header","required":true,"schema":{"type":"string","title":"X-Token"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MessageResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyHeader":[]}]}},"/api/demandes/":{"get":{"tags":["Demandes d’approvisionnement"],"summary":"Lister toutes les demandes","description":"Retourne toutes les demandes d’approvisionnement enregistrées dans le système.","operationId":"get_demandes_api_demandes__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/DemandeApprovisionnementOut"},"type":"array","title":"Response Get Demandes Api Demandes  Get"}}}}},"security":[{"APIKeyHeader":[]}]},"post":{"tags":["Demandes d’approvisionnement"],"summary":"Créer une demande d’approvisionnement","description":"Crée une nouvelle demande d’approvisionnement pour un produit donné dans un magasin.","operationId":"creer_demande_api_demandes__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/DemandeApprovisionnementIn"}}},"required":true},"responses":{"201":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/DemandeApprovisionnementOut"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyHeader":[]}]}},"/api/demandes/traiter/{demande_id}":{"post":{"tags":["Demandes d’approvisionnement"],"summary":"Traiter une demande","description":"Valide une demande d’approvisionnement : met son statut à 'traitée' et met à jour le stock.","operationId":"traiter_demande_api_demandes_traiter__demande_id__post","parameters":[{"name":"demande_id","in":"path","required":true,"schema":{"type":"integer","title":"Demande Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyHeader":[]}]}},"/api/magasins/":{"get":{"tags":["Magasins"],"summary":"Lister tous les magasins","description":"Retourne la liste de tous les magasins enregistrés.","operationId":"get_magasins_api_magasins__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/MagasinDTO"},"type":"array","title":"Response Get Magasins Api Magasins  Get"}}}}},"security":[{"APIKeyHeader":[]}]},"post":{"tags":["Magasins"],"summary":"Créer un magasin","description":"Ajoute un nouveau magasin dans la base de données.","operationId":"post_magasin_api_magasins__post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MagasinCreate"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MagasinDTO"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyHeader":[]}]}},"/api/magasins/{id}/stock":{"get":{"tags":["Magasins"],"summary":"Afficher le stock d’un magasin","description":"Retourne les produits en stock pour un magasin spécifique (via son ID).","operationId":"get_stock_magasin_api_magasins__id__stock_get","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"type":"array","items":{"$ref":"#/components/schemas/ProduitStockDTO"},"title":"Response Get Stock Magasin Api Magasins  Id  Stock Get"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyHeader":[]}]}},"/api/magasins/{id}":{"put":{"tags":["Magasins"],"summary":"Modifier un magasin","description":"Met à jour les informations d’un magasin existant (nom, adresse, etc.).","operationId":"update_magasin_api_magasins__id__put","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"requestBody":{"required":true,"content":{"application/json":{"schema":{"$ref":"#/components/schemas/MagasinCreate"}}}},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/MagasinDTO"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyHeader":[]}]},"delete":{"tags":["Magasins"],"summary":"Supprimer un magasin","description":"Supprime un magasin s’il existe dans la base de données.","operationId":"delete_magasin_api_magasins__id__delete","parameters":[{"name":"id","in":"path","required":true,"schema":{"type":"integer","title":"Id"}}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyHeader":[]}]}},"/api/stock/":{"get":{"tags":["Stock"],"summary":"Lister le stock global","description":"Retourne tous les produits avec leurs quantités en stock dans chaque magasin.","operationId":"get_stock_api_stock__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"items":{"$ref":"#/components/schemas/ProduitStockDTO"},"type":"array","title":"Response Get Stock Api Stock  Get"}}}}},"security":[{"APIKeyHeader":[]}]}},"/api/performance/global":{"get":{"tags":["Performance"],"summary":"Afficher les performances agrégées du système","description":"Retourne les performances globales du système de vente.","operationId":"get_performance_globale_api_performance_global_get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/PerformanceGlobaleDTO"}}}}},"security":[{"APIKeyHeader":[]}]}},"/api/rapports/ventes":{"get":{"tags":["Rapports"],"summary":"Rapport consolidé des ventes","description":"Génère un rapport consolidé des ventes entre deux dates spécifiques.","operationId":"rapport_ventes_api_rapports_ventes_get","parameters":[{"name":"date_debut","in":"query","required":true,"schema":{"type":"string","format":"date","description":"Date de début (YYYY-MM-DD)","title":"Date Debut"},"description":"Date de début (YYYY-MM-DD)"},{"name":"date_fin","in":"query","required":true,"schema":{"type":"string","format":"date","description":"Date de fin (YYYY-MM-DD)","title":"Date Fin"},"description":"Date de fin (YYYY-MM-DD)"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{"$ref":"#/components/schemas/RapportVenteResponse"}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}},"security":[{"APIKeyHeader":[]}]}},"/":{"get":{"summary":"Ping","operationId":"ping__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}},"security":[{"APIKeyHeader":[]}]}}},"components":{"schemas":{"DemandeApprovisionnementIn":{"properties":{"produit_id":{"type":"integer","title":"Produit Id"},"quantite":{"type":"integer","title":"Quantite"},"magasin_id":{"type":"integer","title":"Magasin Id"}},"type":"object","required":["produit_id","quantite","magasin_id"],"title":"DemandeApprovisionnementIn"},"DemandeApprovisionnementOut":{"properties":{"id":{"type":"integer","title":"Id"},"produit_id":{"type":"integer","title":"Produit Id"},"quantite":{"type":"integer","title":"Quantite"},"magasin_id":{"type":"integer","title":"Magasin Id"},"statut":{"type":"string","title":"Statut"}},"type":"object","required":["id","produit_id","quantite","magasin_id","statut"],"title":"DemandeApprovisionnementOut"},"HTTPValidationError":{"properties":{"detail":{"items":{"$ref":"#/components/schemas/ValidationError"},"type":"array","title":"Detail"}},"type":"object","title":"HTTPValidationError"},"MagasinCreate":{"properties":{"nom":{"type":"string","title":"Nom"},"adresse":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Adresse"},"ville":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Ville"}},"type":"object","required":["nom"],"title":"MagasinCreate"},"MagasinDTO":{"properties":{"nom":{"type":"string","title":"Nom"},"adresse":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Adresse"},"ville":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Ville"},"id":{"type":"integer","title":"Id"}},"type":"object","required":["nom","id"],"title":"MagasinDTO"},"MessageResponse":{"properties":{"message":{"type":"string","title":"Message"}},"type":"object","required":["message"],"title":"MessageResponse"},"PerformanceGlobaleDTO":{"properties":{"total_ventes":{"type":"number","title":"Total Ventes"},"produits":{"items":{"$ref":"#/components/schemas/ProduitVendableDTO"},"type":"array","title":"Produits"},"ruptures":{"items":{"$ref":"#/components/schemas/ProduitSimpleDTO"},"type":"array","title":"Ruptures"},"surstocks":{"items":{"$ref":"#/components/schemas/ProduitSimpleDTO"},"type":"array","title":"Surstocks"},"ventes_par_magasin":{"items":{"$ref":"#/components/schemas/VentesParMagasinDTO"},"type":"array","title":"Ventes Par Magasin"},"tendance_journaliere":{"items":{"$ref":"#/components/schemas/TendanceJourDTO"},"type":"array","title":"Tendance Journaliere"}},"type":"object","required":["total_ventes","produits","ruptures","surstocks","ventes_par_magasin","tendance_journaliere"],"title":"PerformanceGlobaleDTO"},"ProduitSimpleDTO":{"properties":{"nom":{"type":"string","title":"Nom"},"categorie":{"type":"string","title":"Categorie"},"stock":{"type":"integer","title":"Stock"}},"type":"object","required":["nom","categorie","stock"],"title":"ProduitSimpleDTO"},"ProduitStockDTO":{"properties":{"id":{"type":"integer","title":"Id"},"nom":{"type":"string","title":"Nom"},"categorie":{"type":"string","title":"Categorie"},"quantite_stock":{"type":"integer","title":"Quantite Stock"},"prix":{"type":"number","title":"Prix"}},"type":"object","required":["id","nom","categorie","quantite_stock","prix"],"title":"ProduitStockDTO"},"ProduitUpdate":{"properties":{"nom":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Nom"},"categorie":{"anyOf":[{"type":"string"},{"type":"null"}],"title":"Categorie"},"prix":{"anyOf":[{"type":"number"},{"type":"null"}],"title":"Prix"},"quantite_stock":{"anyOf":[{"type":"integer"},{"type":"null"}],"title":"Quantite Stock"}},"type":"object","required":["nom","categorie","prix","quantite_stock"],"title":"ProduitUpdate"},"ProduitVendableDTO":{"properties":{"nom":{"type":"string","title":"Nom"},"categorie":{"type":"string","title":"Categorie"},"quantite":{"type":"integer","title":"Quantite"}},"type":"object","required":["nom","categorie","quantite"],"title":"ProduitVendableDTO"},"RapportVenteDTO":{"properties":{"magasin":{"type":"string","title":"Magasin"},"nombre_ventes":{"type":"integer","title":"Nombre Ventes"},"total_ventes":{"type":"number","title":"Total Ventes"}},"type":"object","required":["magasin","nombre_ventes","total_ventes"],"title":"RapportVenteDTO"},"RapportVenteResponse":{"properties":{"rapport":{"items":{"$ref":"#/components/schemas/RapportVenteDTO"},"type":"array","title":"Rapport"}},"type":"object","required":["rapport"],"title":"RapportVenteResponse"},"TendanceJourDTO":{"properties":{"jour":{"type":"string","title":"Jour"},"total":{"type":"number","title":"Total"}},"type":"object","required":["jour","total"],"title":"TendanceJourDTO"},"ValidationError":{"properties":{"loc":{"items":{"anyOf":[{"type":"string"},{"type":"integer"}]},"type":"array","title":"Location"},"msg":{"type":"string","title":"Message"},"type":{"type":"string","title":"Error Type"}},"type":"object","required":["loc","msg","type"],"title":"ValidationError"},"VenteOut":{"properties":{"id":{"type":"integer","title":"Id"},"date":{"type":"string","format":"date-time","title":"Date"},"magasin_id":{"type":"integer","title":"Magasin Id"},"total":{"type":"number","title":"Total"}},"type":"object","required":["id","date","magasin_id","total"],"title":"VenteOut"},"VentesParMagasinDTO":{"properties":{"magasin":{"type":"string","title":"Magasin"},"total":{"type":"number","title":"Total"}},"type":"object","required":["magasin","total"],"title":"VentesParMagasinDTO"}},"securitySchemes":{"APIKeyHeader":{"type":"apiKey","in":"header","name":"x-token"}}}}